@@ -1,32 +1,60 @@
 module github.com/u-root/u-bmc
 
-go 1.16
+go 1.17
 
 require (
 	github.com/beevik/ntp v0.3.0
+	github.com/cleroux/rtc v0.1.1
 	github.com/digitalocean/go-qemu v0.0.0-20210326154740-ac9e0b687001
 	github.com/fullstorydev/grpcurl v1.8.1
 	github.com/golang/protobuf v1.5.2
-	github.com/google/goexpect v0.0.0-20210430020637-ab937bf7fd6f // indirect
 	github.com/grpc-ecosystem/go-grpc-prometheus v1.2.0
-	github.com/insomniacslk/dhcp v0.0.0-20210621130208-1cac67f12b1e // indirect
 	github.com/jhump/protoreflect v1.8.2
 	github.com/jmhodges/clock v0.0.0-20160418191101-880ee4c33548
 	github.com/jpillora/backoff v1.0.0
 	github.com/letsencrypt/pebble v1.0.1
 	github.com/machinebox/progress v0.2.0
-	github.com/matryer/is v1.4.0 // indirect
 	github.com/mdlayher/genetlink v1.0.0
 	github.com/mdlayher/netlink v1.4.1
 	github.com/miekg/dns v1.1.43
 	github.com/prometheus/client_golang v1.11.0
 	github.com/spf13/afero v1.6.0
 	github.com/tarm/serial v0.0.0-20180830185346-98f6abe2eb07
-	github.com/u-root/u-root v0.0.0-20210625045453-236bcb5a057c
+	github.com/u-root/gobusybox/src v0.0.0-20210529132627-adc854ea4425
+	github.com/u-root/u-root v0.0.0-20210903181157-ff1c8497f204
 	github.com/vishvananda/netlink v1.1.1-0.20200221165523-c79a4b7b4066
 	golang.org/x/crypto v0.0.0-20210616213533-5ff15b29337e
 	golang.org/x/net v0.0.0-20210614182718-04defd469f4e
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
 	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c
 	google.golang.org/grpc v1.39.0
+)
+
+require (
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.1.1 // indirect
+	github.com/digitalocean/go-libvirt v0.0.0-20201209184759-e2a69bcd5bd1 // indirect
+	github.com/dustin/go-humanize v1.0.0 // indirect
+	github.com/google/go-cmp v0.5.6 // indirect
+	github.com/google/goexpect v0.0.0-20210430020637-ab937bf7fd6f // indirect
+	github.com/google/goterm v0.0.0-20200907032337-555d40f16ae2 // indirect
+	github.com/josharian/native v0.0.0-20200817173448-b6b71def0850 // indirect
+	github.com/klauspost/compress v1.10.6 // indirect
+	github.com/klauspost/pgzip v1.2.4 // indirect
+	github.com/matryer/is v1.4.0 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
+	github.com/mdlayher/socket v0.0.0-20210307095302-262dc9984e00 // indirect
+	github.com/prometheus/client_model v0.2.0 // indirect
+	github.com/prometheus/common v0.26.0 // indirect
+	github.com/prometheus/procfs v0.6.0 // indirect
+	github.com/u-root/uio v0.0.0-20210528151154-e40b768296a7 // indirect
+	github.com/ulikunitz/xz v0.5.8 // indirect
+	github.com/vishvananda/netns v0.0.0-20210104183010-2eb08e3e575f // indirect
+	golang.org/x/mod v0.4.2 // indirect
+	golang.org/x/text v0.3.6 // indirect
+	golang.org/x/tools v0.1.1 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	google.golang.org/protobuf v1.26.0 // indirect
+	gopkg.in/square/go-jose.v2 v2.1.9 // indirect
 )
