@@ -2,7 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build arm
+//go:build arm
+
 // Assume arm hosts are the BMC
 
 package aspeed
