@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build arm amd64
+//go:build arm || arm64 || amd64
 
 package bmc
 
