@@ -24,6 +24,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/cleroux/rtc"
 	"github.com/golang/protobuf/proto"
 	"github.com/jpillora/backoff"
 	"github.com/prometheus/client_golang/prometheus"
@@ -31,7 +32,6 @@ import (
 	"github.com/u-root/u-bmc/pkg/bmc/cert"
 	"github.com/u-root/u-bmc/pkg/bmc/ttime"
 	pb "github.com/u-root/u-bmc/proto"
-	"github.com/u-root/u-root/pkg/rtc"
 	"golang.org/x/sys/unix"
 )
 
@@ -134,7 +134,7 @@ func startSsh(ak []string) error {
 	// The variable authorizedKeys is generated by Makefiles
 	createFile("/config/authorized_keys", 0600, []byte(strings.Join(ak, "\n")))
 	cmd := exec.Command(
-		"/bbin/sshd",
+		"/bin/sshd",
 		"--ip", "[::0]",
 		"--port", "22",
 		"--privatekey", "/config/ssh_host_ecdsa_key",
@@ -292,13 +292,14 @@ func StartupWithConfig(p Platform, c *config.Config) (error, chan error) {
 			acquireTime(c.RoughtimeServers, c.NtpServers)
 			timeAcquired <- true
 		}
-		r, err := rtc.OpenRTC()
+		r, err := rtc.NewRTC("/dev/rtc0")
 		if err != nil {
 			log.Printf("Failed to open RTC: %v", err)
 			return
 		}
+		defer r.Close()
 		tu := time.Now().UTC()
-		if err := r.Set(tu); err != nil {
+		if err := r.SetTime(tu); err != nil {
 			log.Printf("Failed to update RTC: %v", err)
 		}
 	}()
@@ -413,7 +414,7 @@ func asyncStartup(p Platform, c *config.Config, rpc RPCServer, cm *cert.Manager,
 }
 
 func Shell() {
-	cmd := exec.Command("/bbin/login")
+	cmd := exec.Command("/bin/login")
 	cmd.Dir = "/"
 	cmd.Env = environ
 	cmd.Stdin = os.Stdin
