@@ -1,4 +1,4 @@
-// Copyright 2015-2018 the u-root Authors. All rights reserved
+// Copyright 2015-2021 the u-root Authors. All rights reserved
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
@@ -7,92 +7,99 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
+	"path/filepath"
 
-	"github.com/u-root/u-root/pkg/golang"
-	"github.com/u-root/u-root/pkg/uroot"
-	"github.com/u-root/u-root/pkg/uroot/builder"
-	"github.com/u-root/u-root/pkg/uroot/initramfs"
+	"github.com/u-root/gobusybox/src/pkg/bb"
+	"github.com/u-root/gobusybox/src/pkg/golang"
 )
 
-// Flags for u-root builder.
+// Flags for the gobusybox builder
 var (
-	outputPath = flag.String("o", "", "Path to output initramfs file.")
-	platform   = flag.String("p", "", "Platform to target.")
-	packages   = []string{
-		"github.com/u-root/u-bmc/cmd/fan",
-		"github.com/u-root/u-bmc/cmd/i2cwatcher",
-		"github.com/u-root/u-bmc/cmd/login",
-		"github.com/u-root/u-bmc/cmd/socreset",
-		"github.com/u-root/u-bmc/cmd/ubmcctl",
-		"github.com/u-root/u-root/cmds/core/*",
-	}
+	platform = flag.String("plat", "", "Platform to target.")
 )
 
 func main() {
 	flag.Parse()
 
-	// Main is in a separate functions so defers run on return.
+	// Main is in a separate functions so defers run on return
 	if err := Main(); err != nil {
 		log.Fatal(err)
 	}
-	log.Printf("Successfully wrote initramfs.")
+	log.Print("Successfully created root directory")
 }
 
 // Main is a separate function so defers are run on return, which they wouldn't
-// on exit.
+// on exit
 func Main() error {
+	// Make sure to disable CGO as it's not supported
 	env := golang.Default()
 	if env.CgoEnabled {
-		// TODO(bluecmd): Might need CGO for pcap if that should be included
-		// Given that we already depend on a gcc being available for u-boot and
-		// the linux kernel, this might be fine. Especially if we need to do the
-		// yocto route down the line.
-		log.Printf("Disabling CGO for u-bmc...")
 		env.CgoEnabled = false
 	}
-	log.Printf("Build environment: %s", env)
-	if env.GOOS != "linux" {
-		log.Printf("GOOS is not linux. Did you mean to set GOOS=linux?")
-	}
 
-	builder := builder.BBBuilder{}
-	archiver, err := initramfs.GetArchiver("cpio")
+	// Resolve paths to all individual go commands for the busybox
+	bmccmds, err := filepath.Glob("../cmd/*")
+	if err != nil {
+		return err
+	}
+	urootcmds, err := filepath.Glob("../../u-root/cmds/core/*")
+	if err != nil {
+		return err
+	}
+	platcmds, err := filepath.Glob(fmt.Sprintf("../platform/%s/cmd/*", *platform))
+	if err != nil {
+		return err
+	}
+	var commands []string
+	commands = append(commands, bmccmds...)
+	commands = append(commands, urootcmds...)
+	commands = append(commands, platcmds...)
+	pwd, err := os.Getwd()
 	if err != nil {
 		return err
 	}
 
-	tempDir, err := ioutil.TempDir("", "u-root")
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(tempDir)
-
-	logger := log.New(os.Stderr, "", log.LstdFlags)
-	w, err := archiver.OpenWriter(logger, *outputPath)
-	if err != nil {
-		return err
-	}
-
-	pkgs := append(packages, fmt.Sprintf("github.com/u-root/u-bmc/platform/%s/cmd/*", *platform))
-
-	baseFile := uroot.DefaultRamfs().Reader()
-
-	opts := uroot.Opts{
-		Env: env,
-		Commands: []uroot.Commands{
-			{
-				Builder:  builder,
-				Packages: pkgs,
-			},
+	// Define some options for building the busybox
+	opts := bb.Opts{
+		Env:          env,
+		CommandPaths: commands,
+		BinaryPath:   pwd + "/rootfs/bin/bb",
+		GoBuildOpts: &golang.BuildOpts{
+			NoStrip:        false,
+			NoTrimPath:     false,
+			EnableInlining: false,
 		},
-		BaseArchive:  baseFile,
-		TempDir:      tempDir,
-		OutputFile:   w,
-		InitCmd:      "init",
-		DefaultShell: "elvish",
+		AllowMixedMode: true,
 	}
-	return uroot.CreateInitramfs(logger, opts)
+
+	// Build the bb binary
+	err = bb.BuildBusybox(&opts)
+	if err != nil {
+		return err
+	}
+
+	// Create symlinks for the bb binary and all commands
+	for _, cmd := range commands {
+		err := os.Symlink("bb", "rootfs/bin/"+filepath.Base(cmd))
+		if err != nil {
+			return err
+		}
+	}
+	err = os.Symlink("elvish", "rootfs/bin/sh")
+	if err != nil {
+		return err
+	}
+
+	// Create default directory structure
+	directories := []string{"dev", "proc", "sys", "usr/lib", "var/log", "tmp", "etc", "boot", "config"}
+	for _, dir := range directories {
+		err := os.MkdirAll("rootfs/"+dir, 0755)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
